Refactor: 폴더 네이밍 & 파일 분리
