+++ .storybook/storybook.requires.js
@@ -57,9 +57,9 @@ const getStories = () => {
     "./modules/app/navigation/BottomTabNavigator.stories.tsx": require("../modules/app/navigation/BottomTabNavigator.stories.tsx"),
     "./modules/post/components/button/ConfirmCancelButton.stories.tsx": require("../modules/post/components/button/ConfirmCancelButton.stories.tsx"),
     "./modules/post/components/card/Card.stories.tsx": require("../modules/post/components/card/Card.stories.tsx"),
-    "./modules/post/components/card/display/BottomHeading.stories.tsx": require("../modules/post/components/card/display/BottomHeading.stories.tsx"),
-    "./modules/post/components/card/display/Display.stories.tsx": require("../modules/post/components/card/display/Display.stories.tsx"),
-    "./modules/post/components/card/display/TopHeading.stories.tsx": require("../modules/post/components/card/display/TopHeading.stories.tsx"),
+    "./modules/post/components/card/info/display/Display.stories.tsx": require("../modules/post/components/card/info/display/Display.stories.tsx"),
+    "./modules/post/components/card/info/heading/BottomHeading.stories.tsx": require("../modules/post/components/card/info/heading/BottomHeading.stories.tsx"),
+    "./modules/post/components/card/info/heading/TopHeading.stories.tsx": require("../modules/post/components/card/info/heading/TopHeading.stories.tsx"),
     "./modules/post/components/card/thumbnail/Thumbnail.stories.tsx": require("../modules/post/components/card/thumbnail/Thumbnail.stories.tsx"),
     "./modules/post/components/cardList/CardList.stories.tsx": require("../modules/post/components/cardList/CardList.stories.tsx"),
     "./modules/post/components/creation/meetingNameInput/MeetingNameInput.stories.tsx": require("../modules/post/components/creation/meetingNameInput/MeetingNameInput.stories.tsx"),
