+++ modules/post/components/card/info/display/DecisionDisplay.tsx
