+++ modules/post/components/card/info/display/type.ts
@@ -17,27 +17,3 @@ export interface IGroupDisplay {
 export interface IDecisionDisplay {
   isDecided: boolean;
 }
-
-export interface ITitle {
-  text: string;
-}
-
-interface IRage {
-  startFrom: string;
-  endTo: string;
-}
-
-export interface ISubTitle {
-  userName: string;
-  range: IRage;
-}
-
-export interface IBottomHeading {
-  title: string;
-  subTitle: ISubTitle;
-}
-
-export interface ITopHeading {
-  people: number;
-  isDecided: boolean;
-}
