+++ modules/post/components/card/info/heading/TopHeading.tsx
@@ -1,11 +1,11 @@
 import { View } from 'react-native';
 import React from 'react';
 
-import GroupDisplay from './GroupDisplay';
-import DecisionDisplay from './DecisionDisplay';
-import CardMenu from '../menu/Menu';
-import { ITopHeading } from './type';
+import GroupDisplay from '../display/GroupDisplay';
+import DecisionDisplay from '../display/DecisionDisplay';
+import CardMenu from '../../menu/Menu';
 import useStyles from './style';
+import { ITopHeading } from './type';
 
 export default function TopHeading({ people, isDecided }: ITopHeading) {
   const { cTopHeading, cTopWrap } = useStyles();
