+++ modules/post/components/card/info/heading/style.ts
@@ -1,0 +1,15 @@
+import { makeStyles } from '@rneui/themed';
+
+export default makeStyles(() => ({
+  cTopHeading: {
+    flexDirection: 'row',
+  },
+  cTopWrap: {
+    height: 36,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginVertical: 6,
+    marginHorizontal: 12,
+  },
+}));
