+++ modules/post/components/card/info/heading/type.ts
@@ -1,0 +1,11 @@
+import { IsubTitle } from '../title/type';
+
+export interface IBottomHeading {
+  title: string;
+  subTitle: IsubTitle;
+}
+
+export interface ITopHeading {
+  people: number;
+  isDecided: boolean;
+}
