+++ modules/post/components/card/info/title/SubTitle.tsx
@@ -5,10 +5,10 @@ import Font from '@shared/components/font/Font';
 import Icon from '@shared/components/icon/Icon';
 import { Divider } from '@rneui/themed';
 import { formatDateRange } from '@shared/utils/utils';
-import { ISubTitle } from './type';
 import useStyles from './style';
+import { IsubTitle } from './type';
 
-export default function SubTitle({ userName, range }: ISubTitle) {
+export default function SubTitle({ userName, range }: IsubTitle) {
   const { cSubTitle, subTitleIcon, subTitleFont, divider, dividerBorder } =
     useStyles();
