+++ modules/post/components/card/info/title/Title.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import Font from '@shared/components/font/Font';
 
 import useStyles from './style';
-import { ITitle } from './type';
+import { Ititle } from './type';
 
 /**
  * 모임 게시물 카드 하단 타이틀 컴포넌트
  */
-export default function Title({ text }: ITitle) {
+export default function Title({ text }: Ititle) {
   const { titleFont } = useStyles();
 
   return (
