+++ modules/post/components/card/info/title/type.ts
@@ -1,0 +1,13 @@
+interface Irage {
+  startFrom: string;
+  endTo: string;
+}
+
+export interface IsubTitle {
+  userName: string;
+  range: Irage;
+}
+
+export interface Ititle {
+  text: string;
+}
